Memory utilization optimization
Moved memory allocation that where happening with every read to a kt_forpool
shared thread data structure.
Not thoroughly tested but it seems to be working identically to bwa aln + samse.

BUGFIXES
bwa_seq_t->seq was not being complimented back for bwa_refine_gapped1
This is necessary because of the merging of aln and samse. samse routined no longer
start from freshly loaded reads. It receives the complemented sequences from aln routines.